new v

+++ app/src/main/java/com/silho/ideo/clockwidget/model/openweathermap/Wind.java
@@ -15,7 +15,7 @@ public class Wind implements Parcelable
     private Double speed;
     @SerializedName("deg")
     @Expose
-    private Integer deg;
+    private Double deg;
     public final static Parcelable.Creator<Wind> CREATOR = new Creator<Wind>() {
 
 
@@ -35,7 +35,7 @@ public class Wind implements Parcelable
 
     protected Wind(Parcel in) {
         this.speed = ((Double) in.readValue((Double.class.getClassLoader())));
-        this.deg = ((Integer) in.readValue((Integer.class.getClassLoader())));
+        this.deg = ((Double) in.readValue((Double.class.getClassLoader())));
     }
 
     /**
@@ -50,7 +50,7 @@ public class Wind implements Parcelable
      * @param speed
      * @param deg
      */
-    public Wind(Double speed, Integer deg) {
+    public Wind(Double speed, Double deg) {
         super();
         this.speed = speed;
         this.deg = deg;
@@ -64,11 +64,11 @@ public class Wind implements Parcelable
         this.speed = speed;
     }
 
-    public Integer getDeg() {
+    public Double getDeg() {
         return deg;
     }
 
-    public void setDeg(Integer deg) {
+    public void setDeg(Double deg) {
         this.deg = deg;
     }
